+++ src/main/java/net/detectivekaktus/sound/item/DotcItemSounds.java
@@ -32,6 +32,7 @@ public class DotcItemSounds {
     public static final SoundEvent URN_OF_SHADOWS = DotcSounds.register("item_urn_of_shadows");
     public static final SoundEvent SPIRIT_VESSEL = DotcSounds.register("item_spirit_vessel");
     public static final SoundEvent FORCE_STAFF = DotcSounds.register("item_force_staff");
+    public static final SoundEvent HURRICANE_PIKE = DotcSounds.register("item_hurricane_pike");
 
     public static void initialize() {}
 }